requirements.txt
alembic==1.9.0
attrs==22.1.0
azure-ai-ml==1.2.0
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.12.0
azure-mgmt-core==1.3.2
azure-storage-blob==12.13.0
azure-storage-file-datalake==12.8.0
azure-storage-file-share==12.10.1
azureml-mlflow==1.48.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
databricks-cli==0.17.4
docker==6.0.1
entrypoints==0.4
Flask==2.2.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-auth==2.15.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.2.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.2
mlflow==2.0.1
mlflow-skinny==2.0.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
portalocker==2.6.0
protobuf==4.21.12
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydash==5.1.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.45
sqlparse==0.4.3
strictyaml==1.6.2
tabulate==0.9.0
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
websocket-client==1.4.2
Werkzeug==2.2.2
zipp==3.11.0
